+++ cmd/tetra/commands_windows.go
@@ -1,0 +1,12 @@
+// SPDX-License-Identifier: Apache-2.0
+// Copyright Authors of Tetragon
+
+package main
+
+import (
+	"github.com/spf13/cobra"
+)
+
+func addCommands(rootCmd *cobra.Command) {
+	addBaseCommands(rootCmd)
+}
